+++ CHANGELOG.md
@@ -2,8 +2,17 @@
 most recent version is listed first.
 
 
+
+## **version:** v0.7.6
+- support python version 3.8: https://github.com/komuw/naz/pull/186    
+
+
 ## **version:** v0.7.5
-- support python version 3.8: https://github.com/komuw/naz/pull/186
+- Rename `naz.nazcodec.BaseNazCodec` to `naz.codec.BaseCodec`: https://github.com/komuw/naz/pull/191  
+- Rename `naz.Client.codec_class` to `naz.Client.codec`: https://github.com/komuw/naz/pull/191  
+- Add ability for `naz.log.SimpleLogger` to render logs as a python dictionary: https://github.com/komuw/naz/pull/191  
+- BugFix, `data_coding` should not be applied to all fields that are of type C-octet string: https://github.com/komuw/naz/pull/190     
+- Have the naz messsage protocol only use utf8 for encoding/decoding: https://github.com/komuw/naz/pull/190    
 
 
 ## **version:** v0.7.4